// agent/src/extensions/plannotator/generated/review-core.ts
export type DiffType =
  | "uncommitted"
  | "staged"
  | "unstaged"
  | "last-commit"
  | "jj-current"
  | "jj-last"
  | "jj-line"
  | "jj-all"
  | "branch"
  | "merge-base"
  | "all"
  | `worktree:${string}`
  | "p4-default"
  | `p4-changelist:${string}`;